# Processor cpu optimisations and tracing