--- a/jobsdb/jobsdb.go
+++ b/jobsdb/jobsdb.go
@@ -2462,6 +2462,7 @@ func (jd *HandleT) migrateDSLoop(ctx context.Context) {
 					noJobsMigrated, _ := jd.migrateJobs(ds, migrateTo)
 					totalJobsMigrated += noJobsMigrated
 				}
+				jd.logger.Infof("[[ %s : migrateDSLoop ]]: Total migrated %d jobs", jd.tablePrefix, totalJobsMigrated)
 
 				if totalJobsMigrated <= 0 {
 					jd.dsListLock.Lock()
@@ -2469,6 +2470,7 @@ func (jd *HandleT) migrateDSLoop(ctx context.Context) {
 					jd.inProgressMigrationTargetDS = nil
 					jd.dsListLock.Unlock()
 				}
+				jd.logger.Infof("[[ %s : migrateDSLoop ]]: Migrate DONE", jd.tablePrefix)
 
 				jd.JournalMarkDone(opID)
 			}