--- a/jobsdb/jobsdb.go
+++ b/jobsdb/jobsdb.go
@@ -1628,8 +1628,8 @@ func (jd *HandleT) migrateJobs(srcDS dataSetT, destDS dataSetT) (noJobsMigrated 
 	queryStat := stats.NewTaggedStat("migration_jobs", stats.TimerType, stats.Tags{"customVal": jd.tablePrefix})
 	queryStat.Start()
 	defer queryStat.End()
-	// jd.dsListLock.RLock()
-	// defer jd.dsListLock.RUnlock()
+	jd.dsListLock.RLock()
+	defer jd.dsListLock.RUnlock()
 
 	//Unprocessed jobs
 	unprocessedList := jd.getUnprocessedJobsDS(srcDS, false, 0, GetQueryParamsT{})