--- a/mocks/processor/transformer/mock_transformer.go
+++ b/mocks/processor/transformer/mock_transformer.go
@@ -5,6 +5,7 @@
 package mocks_transformer
 
 import (
+	context "context"
 	reflect "reflect"
 
 	gomock "github.com/golang/mock/gomock"
@@ -47,17 +48,17 @@ func (mr *MockTransformerMockRecorder) Setup() *gomock.Call {
 }
 
 // Transform mocks base method.
-func (m *MockTransformer) Transform(arg0 []transformer.TransformerEventT, arg1 string, arg2 int) transformer.ResponseT {
+func (m *MockTransformer) Transform(arg0 context.Context, arg1 []transformer.TransformerEventT, arg2 string, arg3 int) transformer.ResponseT {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Transform", arg0, arg1, arg2)
+	ret := m.ctrl.Call(m, "Transform", arg0, arg1, arg2, arg3)
 	ret0, _ := ret[0].(transformer.ResponseT)
 	return ret0
 }
 
 // Transform indicates an expected call of Transform.
-func (mr *MockTransformerMockRecorder) Transform(arg0, arg1, arg2 interface{}) *gomock.Call {
+func (mr *MockTransformerMockRecorder) Transform(arg0, arg1, arg2, arg3 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Transform", reflect.TypeOf((*MockTransformer)(nil).Transform), arg0, arg1, arg2)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Transform", reflect.TypeOf((*MockTransformer)(nil).Transform), arg0, arg1, arg2, arg3)
 }
 
 // Validate mocks base method.