--- a/processor/integrations/integrations.go
+++ b/processor/integrations/integrations.go
@@ -1,11 +1,11 @@
 package integrations
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
 	"strings"
 
+	jsoniter "github.com/json-iterator/go"
 	"github.com/rudderlabs/rudder-server/config"
 	backendconfig "github.com/rudderlabs/rudder-server/config/backend-config"
 	"github.com/rudderlabs/rudder-server/services/stats"
@@ -17,6 +17,8 @@ import (
 )
 
 var (
+	jsonfast = jsoniter.ConfigCompatibleWithStandardLibrary
+
 	destTransformURL      string
 	postParametersTFields []string
 )
@@ -83,7 +85,7 @@ type TransResponseT struct {
 
 func CollectDestErrorStats(input []byte) {
 	var integrationStat TransStatsT
-	err := json.Unmarshal(input, &integrationStat)
+	err := jsonfast.Unmarshal(input, &integrationStat)
 	if err == nil {
 		if len(integrationStat.StatTags) > 0 {
 			stats.NewTaggedStat("integration.failure_detailed", stats.CountType, integrationStat.StatTags).Increment()
@@ -93,7 +95,7 @@ func CollectDestErrorStats(input []byte) {
 
 func CollectIntgTransformErrorStats(input []byte) {
 	var integrationStats []TransStatsT
-	err := json.Unmarshal(input, &integrationStats)
+	err := jsonfast.Unmarshal(input, &integrationStats)
 	if err == nil {
 		for _, integrationStat := range integrationStats {
 			if len(integrationStat.StatTags) > 0 {
@@ -106,7 +108,7 @@ func CollectIntgTransformErrorStats(input []byte) {
 
 // GetPostInfo parses the transformer response
 func ValidatePostInfo(transformRawParams PostParametersT) error {
-	transformRaw, err := json.Marshal(transformRawParams)
+	transformRaw, err := jsonfast.Marshal(transformRawParams)
 	if err != nil {
 		return err
 	}