--- a/processor/transformer/transformer.go
+++ b/processor/transformer/transformer.go
@@ -4,13 +4,14 @@ package transformer
 
 import (
 	"bytes"
-	"encoding/json"
 	"fmt"
 	"io"
 	"net/http"
 	"strconv"
 	"sync"
 	"time"
+
+	jsoniter "github.com/json-iterator/go"
 
 	"github.com/rudderlabs/rudder-server/config"
 	backendconfig "github.com/rudderlabs/rudder-server/config/backend-config"
@@ -27,6 +28,8 @@ const (
 	DestTransformerStage        = "dest_transformer"
 	TrackingPlanValidationStage = "trackingPlan_validation"
 )
+
+var jsonfast = jsoniter.ConfigCompatibleWithStandardLibrary
 
 type MetadataT struct {
 	SourceID            string                            `json:"sourceId"`
@@ -284,7 +287,7 @@ func statsTags(event TransformerEventT) stats.Tags {
 
 func (trans *HandleT) request(url string, data []TransformerEventT) []TransformerResponseT {
 	//Call remote transformation
-	rawJSON, err := json.Marshal(data)
+	rawJSON, err := jsonfast.Marshal(data)
 	if err != nil {
 		panic(err)
 	}
@@ -354,7 +357,7 @@ func (trans *HandleT) request(url string, data []TransformerEventT) []Transforme
 	var transformerResponses []TransformerResponseT
 	if resp.StatusCode == http.StatusOK {
 		integrations.CollectIntgTransformErrorStats(respData)
-		err = json.Unmarshal(respData, &transformerResponses)
+		err = jsonfast.Unmarshal(respData, &transformerResponses)
 		//This is returned by our JS engine so should  be parsable
 		//but still handling it
 		if err != nil {