--- a/utils/misc/misc.go
+++ b/utils/misc/misc.go
@@ -29,6 +29,8 @@ import (
 	"time"
 	"unicode"
 
+	jsoniter "github.com/json-iterator/go"
+
 	"github.com/araddon/dateparse"
 	"github.com/bugsnag/bugsnag-go"
 	"github.com/cenkalti/backoff"
@@ -46,6 +48,7 @@ import (
 var AppStartTime int64
 var errorStorePath string
 var reservedFolderPaths []*RFP
+var jsonfast = jsoniter.ConfigCompatibleWithStandardLibrary
 
 const (
 	// RFC3339Milli with milli sec precision
@@ -203,7 +206,7 @@ func GetRudderEventVal(key string, rudderEvent types.SingularEventT) (interface{
 //ParseRudderEventBatch looks for the batch structure inside event
 func ParseRudderEventBatch(eventPayload json.RawMessage) ([]types.SingularEventT, bool) {
 	var gatewayBatchEvent types.GatewayBatchRequestT
-	err := json.Unmarshal(eventPayload, &gatewayBatchEvent)
+	err := jsonfast.Unmarshal(eventPayload, &gatewayBatchEvent)
 	if err != nil {
 		pkgLogger.Debug("json parsing of event payload failed ", string(eventPayload))
 		return nil, false
@@ -512,6 +515,15 @@ func GetIPFromReq(req *http.Request) string {
 }
 
 func ContainsString(slice []string, str string) bool {
+	for _, s := range slice {
+		if s == str {
+			return true
+		}
+	}
+	return false
+}
+
+func ContainsInt(slice []int, str int) bool {
 	for _, s := range slice {
 		if s == str {
 			return true