--- a/utils/misc/uuid_bench_test.go
+++ b/utils/misc/uuid_bench_test.go
@@ -1,0 +1,48 @@
+package misc_test
+
+import (
+	"testing"
+
+	uuid "github.com/gofrs/uuid"
+	"github.com/rudderlabs/rudder-server/utils/misc"
+
+	gluuid "github.com/google/uuid"
+)
+
+var uuidGOOGLE gluuid.UUID
+var uuidGOFRS uuid.UUID
+
+func init() {
+	gluuid.EnableRandPool()
+}
+
+func Benchmark_GOOGLE_UUID(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		uuidGOOGLE = gluuid.New()
+	}
+}
+
+func Benchmark_GOOGLE_UUID_STR_GOFRS(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		uuidGOFRS = uuid.FromStringOrNil(gluuid.New().String())
+	}
+}
+
+func Benchmark_GOOGLE_UUID_BIN_GOFRS(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		b, _ := gluuid.New().MarshalBinary()
+		uuidGOFRS = uuid.FromBytesOrNil(b)
+	}
+}
+
+func Benchmark_FAST_UUID(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		uuidGOFRS = misc.FastUUID()
+	}
+}
+
+func Benchmark_GOFRS_UUID(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		uuidGOFRS = uuid.Must(uuid.NewV4())
+	}
+}