… .enterprise-commit
… go.sum
… processor/processor_test.go