Requires https://github.com/rudderlabs/rudder-server-enterprise/pull/62


# Impact 

It is hard to gauge the exact impact of the changes overall, given that the transformation path has a lot of dependencies, thus creating benchmarks or isolating performance requires a lot of effort.

I run rudder-server under the same load and configuration with and without the changes for a long duration (more than an hour) and measure the average CPU and dest processing times: 

## CPU

**Without** the changes: 
<img width="1506" alt="Screenshot 2022-01-30 at 10 01 16 PM" src="https://user-images.githubusercontent.com/733195/151715651-01d83122-803a-41a8-9440-3ded047c79af.png">

**With** the changes: 
<img width="1160" alt="Screenshot 2022-01-30 at 7 15 22 PM" src="https://user-images.githubusercontent.com/733195/151711135-1cf9cffd-3eaa-4e23-9c65-594a1cf7cb55.png">

Around ~15% less CPU usage is expected.

## Processing

**With** the changes: 
<img width="1484" alt="Screenshot 2022-01-30 at 7 11 59 PM" src="https://user-images.githubusercontent.com/733195/151715688-33c162f2-220d-4e33-a7f4-06e43534e74a.png">

**Without** the changes: 
<img width="1506" alt="Screenshot 2022-01-30 at 9 58 00 PM" src="https://user-images.githubusercontent.com/733195/151715695-56689330-7875-4767-ba34-d90c180423ee.png">

# How 

## Observability

`/profile` and `/trace` endpoints was used extensively to find critical paths needing optimization. In addition to that custom tracing annotation was added, to improve observability over transformation code.

## Optimise JSON (Un) Marshaling

Start using  [json-iterator](https://github.com/json-iterator/go) for (un)marshaling operations. 


## Optimise UUID generation 

by replacing `gofrs/uuid` with  `google/uuid`.

As you can observe from the benchmarks below `google/uuid` is much faster than  `gofrs/uuid`. The main reason is google one is using a random pool so it doesn't have to do a syscall every time.

However, replacing gofrs completely wasn't something I wanted to risk doing just before the new release. Since it would require significant testing. So I've created `misc.FastUUID()`, which generates uuid with google's implementation and coverts it via binary to gofrs struct.

### benchmarks

`"github.com/google/uuid"` vs `"github.com/gofrs/uuid"`
```
goos: darwin
goarch: arm64
pkg: github.com/rudderlabs/rudder-server/processor
Benchmark_GOOGLE_UUID-10              	24369831	        50.32 ns/op	       0 B/op	       0 allocs/op
Benchmark_GOOGLE_UUID_STR_GOFRS-10    	 4280739	       272.5 ns/op	      96 B/op	       2 allocs/op
Benchmark_GOOGLE_UUID_BIN_GOFRS-10    	21232407	        58.20 ns/op	       0 B/op	       0 allocs/op
Benchmark_FAST_UUID-10                	18141252	        59.94 ns/op	       0 B/op	       0 allocs/op
Benchmark_GOFRS_UUID-10               	 3281251	       346.7 ns/op	      16 B/op	       1 allocs/op
```